// src/app/interfaces/user.interface.ts
export interface UserProfile {
  companyName?: string;
  corporateEmail?: string;
  governmentId?: string;
  drivingLicense?: string;
  shortBio?: string;
  hobbies?: string;
  fullName?: string;
  email?: string;
  phone?: string;
  vehicleType?: string;
  vehicleName?: string;
  vehicleRegNumber?: string;
  facilities?: string;
  instructions?: string;
  profileImageUrl?: string;
  createdAt?: string;
}